# back-end/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "My API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/products/{productId}": {
      "get": {
        "tags": [
          "Products"
        ],
        "description": "Gets one product.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200, 500": {
            "description": ""
          }
        }
      }
    },
    "/products/search/category/{categoryId}": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Gets the list of all products by category.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200, 500": {
            "description": ""
          }
        }
      }
    },
    "/products/search/brand/{brandId}": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Gets the list of all products by brand.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "brandId",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200, 500": {
            "description": ""
          }
        }
      }
    },
    "/products/search/{name}": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "Gets the list of all products by search name.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200, 500": {
            "description": ""
          }
        }
      }
    },
    "/products/add": {
      "post": {
        "tags": [
          "Products"
        ],
        "description": "adds one product to database.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
               "$ref": "#/definitions/Product"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 401, 500": {
            "description": ""
          }
        }
      }
    },
    "/products/edit/{productId}": {
      "put": {
        "tags": [
          "Products"
        ],
        "description": "edit one product in the database.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "productId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 401, 500": {
            "description": ""
          }
        }
      }
    },
...
    "/categories/": {
      "get": {
        "tags": [
          "Categories"
        ],
        "description": "Gets the list of all available Categories.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
    "/categories/add": {
      "post": {
        "tags": [
          "Categories"
        ],
        "description": "Creates a new category.",
        "produces": [
          "json"
        ],
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        ],
        "responses": {
          "401": {
            "description": "Unauthorized"
          },
          "200, 404, 500": {
            "description": ""
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Category": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Watch"
        }
      },
      "required": [
        "Name"
      ]
    },
    "Product": {
      "type": "object",
      "properties": {
        "Name": {
          "type": "string",
          "example": "Xmas star"
        },
        "description": {
          "type": "string",
          "example": "Gold"
        },
        "imageUrl": {
          "type": "string",
          "example": "https://www.instagram.com/"
        },
        "unitPrice": {
          "type": "string",
          "example": "15.34e"
        },
        "stock": {
          "type": "string",
          "example": "10"
        },
        "brandId": {
          "type": "string",
          "example": "23"
        },
        "categoryId": {
          "type": "string",
          "example": "11"
        }
      },
      "required": [
        "Name",
        "description",
        "imageUrl",
        "unitPrice",
        "stock",
        "brandId",
        "categoryId"
      ]
    },
...
  }
}